--- tests/base/module_2.py
@@ -1,5 +1,0 @@
-from tests.test_module import TestModule
-
-
-class Module2(TestModule):
-    pass
