--- tests/base/module_3.py
@@ -1,5 +1,0 @@
-from tests.test_module import TestModule
-
-
-class Module3(TestModule):
-    pass
