+++ tests/clazz/base/module_2.py
@@ -1,0 +1,5 @@
+from tests.clazz.test_module import TestModule
+
+
+class Module2(TestModule):
+    pass
