+++ tests/clazz/base/module_3.py
@@ -1,0 +1,5 @@
+from tests.clazz.test_module import TestModule
+
+
+class Module3(TestModule):
+    pass
