+++ tests/clazz/base/packageC/module_c1.py
@@ -1,5 +1,5 @@
 from autoload.decorator import load_config
-from tests.test_module import TestModule
+from tests.clazz.test_module import TestModule
 
 
 @load_config(order=3)
