+++ tests/clazz/packageA/module_a3.py
@@ -1,0 +1,5 @@
+from tests.clazz.test_module import TestModule
+
+
+class ModuleA3(TestModule):
+    pass
