+++ tests/clazz/packageA/packageB/module_b2.py
@@ -1,0 +1,5 @@
+from tests.clazz.test_module import TestModule
+
+
+class ModuleB2(TestModule):
+    pass
