+++ tests/clazz/packageA/packageB/module_b3.py
@@ -1,0 +1,5 @@
+from tests.clazz.test_module import TestModule
+
+
+class ModuleB3(TestModule):
+    pass
