+++ tests/clazz/test_module.py
