--- tests/packageA/module_a1.py
@@ -1,5 +1,0 @@
-from tests.test_module import TestModule
-
-
-class ModuleA1(TestModule):
-    pass
