--- tests/packageA/module_a3.py
@@ -1,5 +1,0 @@
-from tests.test_module import TestModule
-
-
-class ModuleA3(TestModule):
-    pass
