--- tests/packageA/packageB/module_b2.py
@@ -1,5 +1,0 @@
-from tests.test_module import TestModule
-
-
-class ModuleB2(TestModule):
-    pass
