--- tests/packageA/packageB/module_b3.py
@@ -1,5 +1,0 @@
-from tests.test_module import TestModule
-
-
-class ModuleB3(TestModule):
-    pass
